Replace remote options into the separate struct
